Add implementation to pass the insert/remove test
Add implementation to pass traces/trace-01-ops.cmd

This test is for insert and remove, so this patch only
modifies q_insert_head and q_remove_head functions.

Besides the function code to pass the test, also add
error handling for empty or NULL queue, and memory
allocation fail.

In q_insert_head, argument 's' points to the string
to be stored. Currently this function will not return false
if argument 's' is NULL.